feat: finish useGameStore hook implementation
