+++ src/components/Cards/MediumCard/Details/index.tsx
@@ -5,7 +5,6 @@ import { FiPlay } from 'react-icons/fi';
 import { RotatingLines } from 'react-loader-spinner';
 import IGamesApiDTO from '../../../../dtos/apiDTO';
 import useGamesStore from '../../../../hooks/useGamesStore';
-import FireStoreService from '../../../../services/database';
 import truncate from '../../../../utils/trucate';
 import Modal from '../../../Modal';
 import PlatformsIcons from '../../../PlatformsIcons';
@@ -39,71 +38,17 @@ const MediumCardDetails = ({
   const gen = genres!.map(genre => genre.name);
 
   const [modal, setModal] = useState(false);
-  const [loading, setLoading] = useState(true);
-  // const [savedToList, setSavedToList] = useState(false);
 
   const toggleModal = useCallback(() => {
     setModal(prevState => !prevState);
   }, []);
 
-  // const { user } = useAuth();
-
-  const { user, setSavedToList, savedToList, handleStoreGame } = useGamesStore(id!, name)
-
-
-  const filterStoredGames = useCallback(async () => {
-    const gameAlreadySaved = await FireStoreService.getOne(id, user?.id);
-
-    if (gameAlreadySaved[0]) {
-      setSavedToList(true);
-    }
-
-    setLoading(false);
-  }, [id, setSavedToList, user?.id]);
+  const { savedToList, loading, handleStoreGame, filterStoredGames } = useGamesStore(id!, name)
 
   useEffect(() => {
     filterStoredGames();
   }, [filterStoredGames]);
 
-
-  // const handleStoreGame = async () => {
-  //   const data = {
-  //     id,
-  //     name,
-  //   };
-
-  //   const gameAlreadySaved = await FireStoreService.getOne(id, user?.id);
-
-  //   if (!user) {
-  //     ErrorHandler('Sign-in required');
-  //   } else if (gameAlreadySaved[0]) {
-  //     try {
-  //       await FireStoreService.delete(gameAlreadySaved[0], user?.id);
-  //       setSavedToList(false);
-  //       toast.success('Jogo removido com sucesso', {
-  //         style: {
-  //           backgroundColor: '#F65F5F',
-  //           color: '#FFF',
-  //         },
-  //       });
-  //     } catch (e) {
-  //       ErrorHandler(`Oops - ${e}`);
-  //     }
-  //   } else {
-  //     try {
-  //       await FireStoreService.create(data, user!.id);
-  //       setSavedToList(true);
-  //       toast.success('Jogo adicionado com sucesso', {
-  //         style: {
-  //           backgroundColor: '#4316DB',
-  //           color: '#FFF',
-  //         },
-  //       });
-  //     } catch (e) {
-  //       ErrorHandler(`Oops - ${e}`);
-  //     }
-  //   }
-  // };
 
   return (
     <>
